+++ test/configCases/context-modules/exclusion/index.js
