--- test/configCases/context-modules/exclusion/simple/webpack.config.js
@@ -1,6 +1,0 @@
-const webpack = require("../../../../../");
-
-/** @type {import("../../../../../").Configuration} */
-module.exports = {
-	plugins: [new webpack.ContextExclusionPlugin(/dont/)]
-};
