+++ test/configCases/context-modules/exclusion/some-dir/check-here/check-here/file.js
