+++ test/configCases/context-modules/exclusion/some-dir/check-here/dont-check-here/file.js
