+++ test/configCases/context-modules/exclusion/some-dir/check-here/file.js
