+++ test/configCases/context-modules/exclusion/some-dir/dont-check-here/file.js
