+++ test/configCases/context-modules/exclusion/some-dir/dont.js
