+++ test/configCases/context-modules/exclusion/some-dir/file.js
