+++ test/configCases/context-modules/exclusion/webpack.config.js
@@ -1,0 +1,6 @@
+const webpack = require("../../../../");
+
+/** @type {import("../../../../").Configuration} */
+module.exports = {
+	plugins: [new webpack.ContextExclusionPlugin(/dont/)]
+};
