+++ test/configCases/context-modules/replacement-System.import/index.js
