+++ test/configCases/context-modules/replacement-System.import/modules/module-b.js
