+++ test/configCases/context-modules/replacement-a/index.js
