+++ test/configCases/context-modules/replacement-a/webpack.config.js
@@ -1,0 +1,13 @@
+const webpack = require("../../../../");
+
+/** @type {import("../../../../").Configuration} */
+module.exports = {
+	plugins: [
+		new webpack.ContextReplacementPlugin(
+			/replacement.a$/,
+			"new-context",
+			true,
+			/^replaced$/
+		)
+	]
+};
