+++ test/configCases/context-modules/replacement-b/error.js
