+++ test/configCases/context-modules/replacement-b/index.js
