+++ test/configCases/context-modules/replacement-b/only-this.js
