+++ test/configCases/context-modules/replacement-b/webpack.config.js
@@ -1,0 +1,6 @@
+const webpack = require("../../../../");
+
+/** @type {import("../../../../").Configuration} */
+module.exports = {
+	plugins: [new webpack.ContextReplacementPlugin(/replacement.b$/, /^\.\/only/)]
+};
