+++ test/configCases/context-modules/replacement-c/index.js
