+++ test/configCases/context-modules/replacement-c/modules/a.js
