+++ test/configCases/context-modules/replacement-c/modules/module-b.js
