+++ test/configCases/context-modules/replacement-c/webpack.config.js
@@ -1,11 +1,11 @@
 const path = require("path");
-const webpack = require("../../../../../");
+const webpack = require("../../../../");
 
-/** @type {import("../../../../../").Configuration} */
+/** @type {import("../../../../").Configuration} */
 module.exports = {
 	plugins: [
 		new webpack.ContextReplacementPlugin(
-			/context-replacement.c$/,
+			/replacement.c$/,
 			path.resolve(__dirname, "modules"),
 			{
 				a: "./a",
