+++ test/configCases/context-modules/replacement-d/index.js
