+++ test/configCases/context-modules/replacement-d/modules/a.js
