+++ test/configCases/context-modules/replacement-e/index.js
