+++ test/configCases/context-modules/replacement-e/new-context/modules/replaced.js
