+++ test/configCases/context-modules/replacement-e/webpack.config.js
@@ -1,14 +1,14 @@
 const path = require("path");
-const webpack = require("../../../../../");
+const webpack = require("../../../../");
 
-/** @type {import("../../../../../").Configuration} */
+/** @type {import("../../../../").Configuration} */
 module.exports = {
 	resolve: {
 		modules: ["...", path.resolve(__dirname, "new-context/modules")]
 	},
 	plugins: [
 		new webpack.ContextReplacementPlugin(
-			/context-replacement.e$/,
+			/replacement.e$/,
 			"new-context",
 			true,
 			/^replaced$|^\.\/modules\/rep/
