+++ test/configCases/context-modules/replacement-f/folder/a.js
