+++ test/configCases/context-modules/replacement-f/folder/nested/error.js
