+++ test/configCases/context-modules/replacement-f/index.js
