+++ test/configCases/context-modules/replacement-f/webpack.config.js
@@ -1,6 +1,6 @@
-const webpack = require("../../../../../");
+const webpack = require("../../../../");
 
-/** @type {import("../../../../../").Configuration} */
+/** @type {import("../../../../").Configuration} */
 module.exports = {
 	plugins: [new webpack.ContextReplacementPlugin(/folder$/, false, /(a|b)/)]
 };
