--- test/configCases/context-modules/replacement/a/webpack.config.js
@@ -1,13 +1,0 @@
-const webpack = require("../../../../../");
-
-/** @type {import("../../../../../").Configuration} */
-module.exports = {
-	plugins: [
-		new webpack.ContextReplacementPlugin(
-			/context-replacement.a$/,
-			"new-context",
-			true,
-			/^replaced$/
-		)
-	]
-};
