--- test/configCases/context-modules/replacement/b/webpack.config.js
@@ -1,8 +1,0 @@
-const webpack = require("../../../../../");
-
-/** @type {import("../../../../../").Configuration} */
-module.exports = {
-	plugins: [
-		new webpack.ContextReplacementPlugin(/context-replacement.b$/, /^\.\/only/)
-	]
-};
